+++ vantage6-algorithm-store/vantage6/algorithm/store/resource/__init__.py
@@ -1,7 +1,10 @@
+from __future__ import annotations
+
 import logging
 from collections.abc import Callable
 from functools import wraps
 from http import HTTPStatus
+from typing import TYPE_CHECKING
 
 from flask import g, request
 from flask_jwt_extended import get_jwt_identity, verify_jwt_in_request
@@ -13,11 +16,13 @@ from vantage6.common.enum import AlgorithmViewPolicies, StorePolicies
 
 from vantage6.backend.common.services_resources import BaseServicesResources
 
-from vantage6.algorithm.store import PermissionManager
 from vantage6.algorithm.store.model.common.enums import DefaultStorePolicies
 from vantage6.algorithm.store.model.policy import Policy
 from vantage6.algorithm.store.model.rule import Operation
 from vantage6.algorithm.store.model.user import User
+
+if TYPE_CHECKING:
+    from vantage6.algorithm.store.permission import PermissionManager
 
 log = logging.getLogger(logger_name(__name__))